refactor(#27): ChatMessageDto를 global로 이동하여 의존 방향 정상화
