+++ src/main/java/io/github/ascrew/monomatbe/global/redis/RedisPublisher.java
@@ -7,7 +7,7 @@
  */
 package io.github.ascrew.monomatbe.global.redis;
 
-import io.github.ascrew.monomatbe.domain.chat.dto.ChatMessageDto;
+import io.github.ascrew.monomatbe.global.websocket.dto.ChatMessageDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
